+++ tests/aux_search/test_main.py
@@ -20,28 +20,27 @@ def test_aux_navbar(aux_page: Page) -> None:
     nav_bar = page.get_by_test_id("aux-nav-bar")
     page.screenshot(path="tests_aux_search_test_main-test_aux_navbar.png")
     expect(nav_bar).to_be_visible()
-    nav_item = nav_bar.get_by_text("Wikidata")
-    expect(nav_item).to_be_visible()
 
 
 @pytest.mark.integration
 @pytest.mark.external
-def test_search_and_import_results(aux_page: Page) -> None:
+def test_wikidata_search_and_import_results(aux_page: Page) -> None:
     page = aux_page
     search_input = page.get_by_placeholder("Search here...")
     expect(search_input).to_be_visible()
 
     # test search input is showing correctly
     search_input.fill("this doesn't exist gs871s9j91kksWN0shx345jnj")
-    expect(page.get_by_text("showing 0 of 0 items found")).to_be_visible()
     page.screenshot(path="tests_aux_search_test_main-test_search-input-0-found.png")
+    expect(page.get_by_text("showing 0 of")).to_be_visible()
 
     # test expand button works
     search_input.fill("rki")
     expand_all_properties_button = page.get_by_test_id("expand-properties-button").nth(
         1
     )
-    expect(page.get_by_text("Robert Koch-Institut")).to_be_visible()
+    page.screenshot(path="tests_aux_search_test_main-search_result.png")
+    expect(page.get_by_text("Robert Koch-Institut").nth(1)).to_be_visible()
     expect(page.get_by_test_id("all-properties-display")).not_to_be_visible()
     expand_all_properties_button.click()
     expect(page.get_by_test_id("all-properties-display")).to_be_visible()
@@ -57,6 +56,47 @@ def test_search_and_import_results(aux_page: Page) -> None:
     # test node was ingested into backend
     connector = BackendApiConnector.get()
     result = connector.fetch_extracted_items(q="Robert Koch-Institut")
+    assert result["total"] >= 1
+
+
+@pytest.mark.integration
+@pytest.mark.external
+def test_ldap_search_and_import_results(aux_page: Page) -> None:
+    page = aux_page
+    ldap_tab = page.get_by_role("tab", name="LDAP")
+    ldap_tab.click()
+    search_input = page.get_by_placeholder("Search here...")
+    expect(search_input).to_be_visible()
+
+    # test search input is showing correctly
+    search_input.fill("doesn't exist gs871s9j91k*")
+    expect(page.get_by_text("Showing 0 of")).to_be_visible()
+    page.screenshot(
+        path="tests_aux_search_test_main-test_ldap_search-input-0-found.png"
+    )
+
+    # test expand button works
+    search_input.fill("Ciftci*")
+    expand_all_properties_button = page.get_by_test_id("expand-properties-button").nth(
+        1
+    )
+    page.screenshot(path="tests_aux_search_test_main-search_result_lap.png")
+    expect(page.get_by_text("Ciftci")).to_be_visible()
+    expect(page.get_by_test_id("all-properties-display")).not_to_be_visible()
+    expand_all_properties_button.click()
+    expect(page.get_by_test_id("all-properties-display")).to_be_visible()
+    page.screenshot(path="tests_aux_search_test_main-test_ldap_expand_button.png")
+
+    # test import button works
+    import_button = page.get_by_text("Import").nth(1)
+    import_button.click()
+    expect(page.get_by_text("Aux search result imported successfully")).to_be_visible()
+    expect(import_button).to_be_disabled()
+    page.screenshot(path="tests_aux_search_test_main-test_ldap_import_button.png")
+
+    # test node was ingested into backend
+    connector = BackendApiConnector.get()
+    result = connector.fetch_extracted_items(q="Ciftci")
     assert result["total"] >= 1